+++ test/scenic/cache/file_test.exs
@@ -95,6 +95,14 @@ defmodule Scenic.Cache.FileReadTest do
              hash: :sha256
            ) == {:error, :enoent}
   end
+
+
+  test "read uses a passed in parser" do
+    # missing term file
+    assert Cache.File.read(@sample_path, :insecure, parser: fn(_) ->
+      :parsed_data
+    end) == :parsed_data
+  end
 end
 
 defmodule Scenic.Cache.FileLoadTest do